worker_scripts/requirements.txt
mysql-connector # TODO Version freeze